Initial commit.
This is a JavaFX application that communicates with an SQLServer database using the JDBC API. The application is built to be used by a store chain that sells printers and parts for printers.

It features 3 menus. The first menu is for the HR department that enables users of the app to view, add, alter and delete employees and their data. The second menu is for the sales department. This menu allows users to view orders, sales, current stock and it also allows them to make sales (if the item is in stock). The third menu is a financial menu. It has reports of financial activity (sales, costs and profit) in the last 30 days.